Move queue to be default entry ponit and old entry point to be legacy, fixes double adding. #108,#111,#112,#113.
